+++ src/reservationPopup/reservationPopupStyles.css
@@ -34,7 +34,7 @@
     height: 280px;
     width: 180px;
     border-radius: 20px;
-}
+    box-shadow: rgba(0, 0, 0, 0.09) 0px 3px 2px, rgba(0, 0, 0, 0.09) 0px 6px 3px, rgba(0, 0, 0, 0.09) 0px 12px 6px, rgba(0, 0, 0, 0.09) 0px 16px 8px, rgba(0, 0, 0, 0.09) 0px 32px 16px;}
 
 .reservation-form {
     display: flex;
@@ -121,6 +121,8 @@
 }
 
 input { 
+    min-width: 190px;
+    width: 190px;
     height: 30px;
     align-self: start;
     box-shadow: rgba(0, 0, 0, 0.19) 0px 10px 20px, rgba(0, 0, 0, 0.23) 0px 6px 6px;
@@ -141,10 +143,12 @@ input {
 .reservation-list {
     position: relative;
     top: -55px;
-    height: 280px;
+    min-height: 280px;
+    max-height: 280px;
     width: 180px;
-    border: 1px solid;
+    border: 1px;
     border-radius: 10px;
+    box-shadow: rgba(0, 0, 0, 0.15) 0px 15px 25px, rgba(0, 0, 0, 0.05) 0px 5px 10px;
 }
 
 .reservation-list-title {
